WPLCT/requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.24
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
filterpy==1.4.5
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.6.1
imutils==0.5.4
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
norfair==0.1.8
numpy==1.19.5
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.2.54
pandas==1.1.5
Pillow @ file:///C:/ci/pillow_1625663293114/work
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
Pygments==2.9.0
pymediainfo==5.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzbar==0.1.8
requests==2.26.0
requests-oauthlib==1.3.0
rich==6.2.0
rsa==4.7.2
scipy==1.5.4
seaborn==0.11.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
tkVideo==0.1
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
yolov5==5.0.7
zipp==3.5.0
